requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
dj-database-url==1.0.0
Django==4.1.2
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.4
pytz==2022.4
sqlparse==0.4.3
